pyproject.toml
[project]
name = "vehicle-detection"
version = "0.1.0"
description = "Detection (localisation & classification) of vehicles on images."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "fire>=0.7.1",
    "gdown>=5.2.0",
    "hydra-core>=1.3.2",
    "lightning>=2.6.0",
    "mlflow>=3.8.1",
    "numpy>=2.3.5",
    "pathlib>=1.0.1",
    "pycocotools>=2.0.11",
    "torch>=2.9.1",
    "torchmetrics>=1.8.2",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.235",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 100
format = { docstring-code-format = true }

[tool.ruff.lint]
select = [
    # isort
    "I",
    # pycodestyle
    "E", "W",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # flake8-builtins
    "A",
    # flake8-return
    "RET",
    # flake8-use-pathlib
    "PTH",
    # Ruff-specific rules
    "RUF",
]

pycodestyle = { max-doc-length = 105 }
